Refactored the AssetController testcase
