+++ application/src/it/java/com/hashmapinc/server/NoSqlIntegrationTestSuite.java
@@ -59,6 +59,7 @@ public class NoSqlIntegrationTestSuite {
         dataSets.add(new CustomCassandraCQLUnit.NamedDataset("2.cql", new ClassPathCQLDataSet("cassandra/upgrade/2.cql" , false, false)));
         dataSets.add(new CustomCassandraCQLUnit.NamedDataset("3.cql", new ClassPathCQLDataSet("cassandra/upgrade/3.cql" , false, false)));
         dataSets.add(new CustomCassandraCQLUnit.NamedDataset("4.cql", new ClassPathCQLDataSet("cassandra/upgrade/4.cql" , false, false)));
+        dataSets.add(new CustomCassandraCQLUnit.NamedDataset("5.cql", new ClassPathCQLDataSet("cassandra/upgrade/5.cql" , false, false)));
         return dataSets;
     }
